Building redux from scratch
This is a new module where we can check the redux library from scratch to know the bahavior and how is working inside. This is a good practice and manage in a better way redux applications. These files describe the principle of redux with some comments on the source code lines.